chore: resize util refactoring
